+++ __tests__/services/connection.service.spec.ts
@@ -10,6 +10,7 @@ import { LoadSceneResponse } from '@proto/world-engine_pb';
 import { v4 } from 'uuid';
 
 import { protoTimestamp } from '../../src/common/helpers';
+import { Logger } from '../../src/common/logger';
 import { InworldPacket } from '../../src/entities/inworld_packet.entity';
 import { Scene } from '../../src/entities/scene.entity';
 import { Session } from '../../src/entities/session.entity';
@@ -57,6 +58,7 @@ test('should return event factory', () => {
       key: KEY,
       secret: SECRET,
     },
+    onError,
   });
 
   expect(connection.getEventFactory()).toBeInstanceOf(EventFactory);
@@ -68,6 +70,7 @@ test('should generate session token', async () => {
       key: KEY,
       secret: SECRET,
     },
+    onError,
   });
   const generateSessionToken = jest
     .spyOn(connection, 'generateSessionToken')
@@ -210,28 +213,6 @@ describe('open', () => {
     expect(onError).toHaveBeenCalledWith(err);
   });
 
-  test('should catch error on load scene and pass it to console.error', async () => {
-    const connection = new ConnectionService({
-      apiKey: { key: KEY, secret: SECRET },
-      name: SCENE,
-      config: { capabilities },
-      user,
-      onMessage,
-      onDisconnect,
-    });
-
-    const err = new Error();
-    const log = jest.spyOn(console, 'error').mockImplementation(() => {});
-    jest
-      .spyOn(connection, 'generateSessionToken')
-      .mockImplementationOnce(() => Promise.reject(err));
-
-    await connection.open();
-
-    expect(log).toHaveBeenCalledTimes(1);
-    expect(log).toBeCalledWith(err);
-  });
-
   test('should catch error on connection establishing and pass it to handler', async () => {
     jest
       .spyOn(connection, 'generateSessionToken')
@@ -280,6 +261,7 @@ describe('open', () => {
       name: SCENE,
       config: { capabilities },
       user,
+      onError,
       onMessage,
     });
 
@@ -468,6 +450,7 @@ describe('close', () => {
   test('should skip for empty stream', () => {
     const connection = new ConnectionService({
       apiKey: { key: KEY, secret: SECRET },
+      onError,
     });
     const end = jest
       .spyOn(ClientDuplexStreamImpl.prototype, 'end')
@@ -479,6 +462,7 @@ describe('close', () => {
   });
 
   test('should execute for existing stream', async () => {
+    const loggerDebug = jest.spyOn(Logger.prototype, 'debug');
     const connection = new ConnectionService({
       apiKey: { key: KEY, secret: SECRET },
       name: SCENE,
@@ -506,10 +490,13 @@ describe('close', () => {
     connection.close();
 
     expect(cancel).toHaveBeenCalledTimes(1);
+    expect(loggerDebug).toHaveBeenCalledTimes(1);
   });
 });
 
 describe('send', () => {
+  const loggerDebug = jest.spyOn(Logger.prototype, 'debug');
+
   beforeEach(() => {
     jest.clearAllMocks();
   });
@@ -517,6 +504,7 @@ describe('send', () => {
   test('should skip for empty stream', async () => {
     const connection = new ConnectionService({
       apiKey: { key: KEY, secret: SECRET },
+      onError,
     });
     const write = jest
       .spyOn(ClientDuplexStreamImpl.prototype, 'write')
@@ -557,6 +545,7 @@ describe('send', () => {
 
     expect(open).toHaveBeenCalledTimes(1);
     expect(write).toHaveBeenCalledTimes(1);
+    expect(loggerDebug).toHaveBeenCalledTimes(1);
   });
 
   test('should throw error in case of connection is inactive on send call', async () => {