+++ src/clients/inworld.client.ts
@@ -19,6 +19,7 @@ import {
   InternalClientConfiguration,
   User,
 } from '../common/data_structures';
+import { Logger } from '../common/logger';
 import { InworldPacket } from '../entities/inworld_packet.entity';
 import { Session } from '../entities/session.entity';
 import { ConnectionService } from '../services/connection.service';
@@ -41,6 +42,15 @@ export class InworldClient<
   private onMessage: ((message: InworldPacketT) => Awaitable<void>) | undefined;
 
   private extension: Extension<InworldPacketT>;
+
+  private logger = Logger.getInstance();
+
+  constructor() {
+    this.onError = (err: ServiceError) => {
+      this.logError(err);
+      console.error(err);
+    };
+  }
 
   setApiKey(apiKey: ApiKey) {
     this.apiKey = apiKey;
@@ -79,7 +89,10 @@ export class InworldClient<
   }
 
   setOnError(fn: (err: ServiceError) => void) {
-    this.onError = fn;
+    this.onError = (err: ServiceError) => {
+      this.logError(err);
+      fn(err);
+    };
 
     return this;
   }
@@ -108,6 +121,7 @@ export class InworldClient<
     return new ConnectionService({
       apiKey: this.apiKey,
       name: this.scene,
+      onError: this.onError,
     }).generateSessionToken();
   }
 
@@ -141,7 +155,7 @@ export class InworldClient<
     const { connection = {}, capabilities = {} } = this.config;
 
     return {
-      ...connection,
+      connection,
       capabilities: this.buildCapabilities(capabilities),
     };
   }
@@ -186,6 +200,10 @@ export class InworldClient<
 
     this.validateScene();
   }
+
+  private logError = (err: ServiceError) => {
+    this.logger.error(err);
+  };
 }
 
 InworldClient.prototype.setGenerateSessionToken = util.deprecate(